Experimental client performance improvement

+++ Shared/ServerPackets.cs
@@ -4974,13 +4974,13 @@ namespace ServerPackets
     {
         public override short Index { get { return (short)ServerPacketIds.NPCImageUpdate; } }
 
-        public long ObjectID;
+        public uint ObjectID;
         public ushort Image;
         public Color Colour;
 
         protected override void ReadPacket(BinaryReader reader)
         {
-            ObjectID = reader.ReadInt64();
+            ObjectID = reader.ReadUInt32();
             Image = reader.ReadUInt16();
             Colour = Color.FromArgb(reader.ReadInt32());
         }
@@ -4995,13 +4995,13 @@ namespace ServerPackets
     {
         public override short Index { get { return (short)ServerPacketIds.MountUpdate; } }
 
-        public long ObjectID;
+        public uint ObjectID;
         public short MountType;
         public bool RidingMount;
 
         protected override void ReadPacket(BinaryReader reader)
         {
-            ObjectID = reader.ReadInt64();
+            ObjectID = reader.ReadUInt32();
             MountType = reader.ReadInt16();
             RidingMount = reader.ReadBoolean();
         }
@@ -5017,12 +5017,12 @@ namespace ServerPackets
     {
         public override short Index { get { return (short)ServerPacketIds.TransformUpdate; } }
 
-        public long ObjectID;
+        public uint ObjectID;
         public short TransformType;
 
         protected override void ReadPacket(BinaryReader reader)
         {
-            ObjectID = reader.ReadInt64();
+            ObjectID = reader.ReadUInt32();
             TransformType = reader.ReadInt16();
         }
         protected override void WritePacket(BinaryWriter writer)
@@ -5068,7 +5068,7 @@ namespace ServerPackets
     {
         public override short Index { get { return (short)ServerPacketIds.FishingUpdate; } }
 
-        public long ObjectID;
+        public uint ObjectID;
         public bool Fishing;
         public int ProgressPercent;
         public int ChancePercent;
@@ -5077,7 +5077,7 @@ namespace ServerPackets
 
         protected override void ReadPacket(BinaryReader reader)
         {
-            ObjectID = reader.ReadInt64();
+            ObjectID = reader.ReadUInt32();
             Fishing = reader.ReadBoolean();
             ProgressPercent = reader.ReadInt32();
             ChancePercent = reader.ReadInt32();
